+++ kopy-compiler/main/kotlin/com/javiersc/kotlin/kopy/compiler/utils/name.kt
@@ -1,5 +1,0 @@
-package com.javiersc.kotlin.kopy.compiler.utils
-
-import org.jetbrains.kotlin.name.Name
-
-internal fun String.toName(): Name = Name.identifier(this)